:recycle: refoctoring du code. Suppression de classes en doublon. Utilisation de l'annotation @Enumerated dans les entités order et client. Suppression des méthodes de conversion enum int en string
