+++ src/main/java/com/tp/crm/controller/OrderController.java
@@ -1,35 +1,38 @@
 package com.tp.crm.controller;
 
-import com.tp.crm.model.*;
-import com.tp.crm.model.dto.OrderPutDto;
+import com.tp.crm.model.dto.OrderPutDTO;
+import com.tp.crm.model.dto.OrderDTO;
+import com.tp.crm.model.dto.OrderPostDTO;
+import com.tp.crm.model.dto.mapper.ClientMapper;
+import com.tp.crm.model.entity.Order;
 import com.tp.crm.service.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @RestController
+@RequestMapping("orders")
 public class OrderController {
 
     @Autowired
     private OrderService orderService;
 
-    @GetMapping("orders/{id}")
+    @GetMapping("{id}")
     public ResponseEntity<OrderDTO> findOrderById(@PathVariable Integer id) {
         Optional<Order> optional = orderService.getOrder(id);
         if(optional.isPresent()) {
             Order order = optional.get();
-            OrderDTO dto = OrderMapper.entityToDto(order);
+            OrderDTO dto = ClientMapper.OrderMapper.entityToDto(order);
             return ResponseEntity.ok(dto);
         } else {
             return ResponseEntity.notFound().build();
         }
     }
-    @PostMapping("orders")
+    @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<OrderPostDTO> addOrder(@RequestBody OrderPostDTO newOrder) {
         OrderPostDTO orderPostDTO = orderService.addOrder(newOrder);
@@ -37,8 +40,8 @@ public class OrderController {
     }
 
 
-    @PutMapping("order/{id}")
-    public ResponseEntity<?> putOrder(@RequestBody OrderPutDto newdata, @PathVariable Integer id) {
+    @PutMapping("{id}")
+    public ResponseEntity<?> putOrder(@RequestBody OrderPutDTO newdata, @PathVariable Integer id) {
 
         if (orderService.notFound(newdata, id)) {
             return ResponseEntity.status(404).body("L'id de l'url est différente de celle envoyer dans le body");
@@ -53,12 +56,12 @@ public class OrderController {
         }
     }
 
-    @GetMapping("orders")
+    @GetMapping
     public List<Order> getOrder(){
         return orderService.getOrders();
     }
 
-    @DeleteMapping("orders/{id}")
+    @DeleteMapping("{id}")
     public ResponseEntity<?> deleteOrderById(@PathVariable Integer id) {
 
         if (orderService.deleteOrderById(id) == null) {
