--- src/main/java/com/tp/crm/model/StatClient.java
@@ -1,6 +1,0 @@
-package com.tp.crm.model;
-
-public enum StatClient {
-    ACTIVE,
-    INACTIVE
-}
