--- src/main/java/com/tp/crm/model/StatOrder.java
@@ -1,7 +1,0 @@
-package com.tp.crm.model;
-
-public enum StatOrder {
-    CANCELED,
-    OPTION,
-    CONFIRMED
-}
