+++ src/main/java/com/tp/crm/model/dto/mapper/OrderMapper.java
@@ -1,12 +1,11 @@
 package com.tp.crm.model.dto.mapper;
 
-import com.tp.crm.model.Order;
-import com.tp.crm.model.dto.OrderPutDto;
-import jakarta.persistence.Entity;
+import com.tp.crm.model.entity.Order;
+import com.tp.crm.model.dto.OrderPutDTO;
 
 public class OrderMapper {
-    public static OrderPutDto entityToDto(Order entity){
-        OrderPutDto orderDto = new OrderPutDto();
+    public static OrderPutDTO entityToDto(Order entity){
+        OrderPutDTO orderDto = new OrderPutDTO();
 
         orderDto.setIdClient(entity.getClient().getId());
         orderDto.setDesignation(entity.getDesignation());
@@ -19,7 +18,7 @@ public class OrderMapper {
         return orderDto;
     }
 
-    public static Order DtoToEntity(OrderPutDto dto){
+    public static Order DtoToEntity(OrderPutDTO dto){
         Order entity = new Order();
         entity.setId(dto.getId());
         entity.setUnitPrice(dto.getUnitPrice());
