+++ src/main/java/com/tp/crm/repository/ClientRepository.java
@@ -1,6 +1,6 @@
 package com.tp.crm.repository;
 
-import com.tp.crm.model.Client;
+import com.tp.crm.model.entity.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
