+++ src/main/java/com/tp/crm/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package com.tp.crm.repository;
 
-import com.tp.crm.model.Order;
+import com.tp.crm.model.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
