+++ src/main/java/com/tp/crm/service/OrderService.java
@@ -1,16 +1,14 @@
 package com.tp.crm.service;
 
-import com.tp.crm.model.Client;
-import com.tp.crm.model.Order;
-import com.tp.crm.model.StatClient;
-import com.tp.crm.model.StatOrder;
-import com.tp.crm.model.OrderPostDTO;
-import com.tp.crm.model.OrderPostMapper;
+import com.tp.crm.model.StateOrder;
+import com.tp.crm.model.entity.Client;
+import com.tp.crm.model.entity.Order;
+import com.tp.crm.model.dto.OrderPostDTO;
+import com.tp.crm.model.dto.mapper.OrderPostMapper;
 import com.tp.crm.repository.ClientRepository;
 import com.tp.crm.repository.OrderRepository;
-import org.springframework.web.bind.annotation.GetMapping;
 import com.tp.crm.model.dto.mapper.OrderMapper;
-import com.tp.crm.model.dto.OrderPutDto;
+import com.tp.crm.model.dto.OrderPutDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -44,15 +42,10 @@ public class OrderService {
         return orderPostDTO;
     }
 
-    // Modification d'une prestation
-    public Order putOrder(OrderPutDto newdata, Integer id) {
+    public Order putOrder(OrderPutDTO newdata, Integer id) {
 
         if (findById(id).isPresent()) {
-            StatOrder etatOrderDto = newdata.getState();
             Order order = OrderMapper.DtoToEntity(newdata);
-            // Gérer l'état de l'order du dto (STRING) en INTEGER
-            gererEtatOrder(etatOrderDto, order);
-            // Envoyer l'objet client en fonction de l'attribut idClient récuperer dans OrderDto
             Client client = clientService.findById(newdata.getIdClient());
             if (client != null)
                 order.setClient(client);
@@ -62,7 +55,7 @@ public class OrderService {
         return null;
     }
 
-    public boolean notFound(OrderPutDto newdata, Integer id) {
+    public boolean notFound(OrderPutDTO newdata, Integer id) {
         if (findById(newdata.getId()).isPresent()) {
             Order newOrder = findById(newdata.getId()).get();
             if (!newOrder.getId().equals(id)) {
@@ -74,22 +67,6 @@ public class OrderService {
 
     public Optional<Order> findById(Integer id) {
         return orderRepository.findById(id);
-    }
-
-    public void gererEtatOrder(StatOrder etat, Order order) {
-        switch (etat) {
-            case CANCELED:
-                order.setState(0);
-                break;
-            case OPTION:
-                order.setState(1);
-                break;
-            case CONFIRMED:
-                order.setState(2);
-                break;
-            default:
-                break;
-        }
     }
